+++ src/IDBIndex.js
@@ -50,6 +50,7 @@ IDBIndex.__createIndex = function (store, index) {
     const columnExists = !!store.__indexes[index.name] && store.__indexes[index.name].__deleted;
 
     // Add the index to the IDBObjectStore
+    index.__pending = true;
     store.__indexes[index.name] = index;
     store.indexNames.push(index.name);
 
@@ -83,6 +84,7 @@ IDBIndex.__createIndex = function (store, index) {
                                 addIndexEntry(i + 1);
                             }
                         } else {
+                            delete index.__pending;
                             success(store);
                         }
                     }
@@ -99,7 +101,7 @@ IDBIndex.__createIndex = function (store, index) {
             CFG.DEBUG && console.log(sql);
             tx.executeSql(sql, [], applyIndex, error);
         }
-    });
+    }, undefined, store);
 };
 
 /**
@@ -122,7 +124,7 @@ IDBIndex.__deleteIndex = function (store, index) {
 
         // Update the object store's index list
         IDBIndex.__updateIndexList(store, tx, success, error);
-    });
+    }, undefined, store);
 };
 
 /**
@@ -177,7 +179,7 @@ IDBIndex.prototype.__fetchIndexData = function (key, opType) {
 
     return me.objectStore.transaction.__addToTransactionQueue(function (...args) {
         fetchIndexData(me, hasKey, encodedKey, opType, ...args);
-    });
+    }, undefined, me);
 };
 
 /**
@@ -187,7 +189,7 @@ IDBIndex.prototype.__fetchIndexData = function (key, opType) {
  * @returns {IDBRequest}
  */
 IDBIndex.prototype.openCursor = function (range, direction) {
-    return new IDBCursor(range, direction, this.objectStore, this, '_' + this.name, 'value').__req;
+    return new IDBCursorWithValue(range, direction, this.objectStore, this, '_' + this.name, 'value').__req;
 };
 
 /**
@@ -197,7 +199,7 @@ IDBIndex.prototype.openCursor = function (range, direction) {
  * @returns {IDBRequest}
  */
 IDBIndex.prototype.openKeyCursor = function (range, direction) {
-    return new IDBCursorWithValue(range, direction, this.objectStore, this, '_' + this.name, 'key').__req;
+    return new IDBCursor(range, direction, this.objectStore, this, '_' + this.name, 'key').__req;
 };
 
 IDBIndex.prototype.get = function (key) {
@@ -222,9 +224,13 @@ IDBIndex.prototype.count = function (key) {
         return this.__fetchIndexData('count');
     }
     if (util.instanceOf(key, IDBKeyRange)) {
-        return new IDBCursor(key, 'next', this.objectStore, this, '_' + this.name, 'value', true).__req;
+        return new IDBCursorWithValue(key, 'next', this.objectStore, this, '_' + this.name, 'value', true).__req;
     }
     return this.__fetchIndexData(key, 'count');
+};
+
+IDBIndex.prototype.toString = function () {
+    return '[object IDBIndex]';
 };
 
 Object.defineProperty(IDBIndex, Symbol.hasInstance, {