+++ tests-polyfill/test-helper.js
@@ -1,7 +1,10 @@
 GLOBAL.window = GLOBAL;
 
-const setGlobalVars = require('../dist/indexeddbshim-node.min');
+const setGlobalVars = require('../dist/indexeddbshim-node');
+// const setGlobalVars = require('../dist/indexeddbshim-node.min');
+
 setGlobalVars();
+// shimIndexedDB.__debug(true);
 
 window.DOMException = indexedDB.modules.DOMException;